third_party/WebKit/LayoutTests/platform/win7/fast/multicol/caret-range-outside-columns-rtl-expected.txt

Test that hitting areas outside columns takes us to the right text content inside columns.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Above first column:
PASS document.caretRangeFromPoint(450,0).startOffset is 0
PASS document.caretRangeFromPoint(350,0).startOffset is 0
PASS document.caretRangeFromPoint(291,0).startOffset is 0
Above second column:
FAIL document.caretRangeFromPoint(290,0).startOffset should be 12. Was 8.
FAIL document.caretRangeFromPoint(250,0).startOffset should be 12. Was 8.
FAIL document.caretRangeFromPoint(191,0).startOffset should be 12. Was 8.
Above third column:
PASS document.caretRangeFromPoint(190,0).startOffset is 24
PASS document.caretRangeFromPoint(150,0).startOffset is 24
PASS document.caretRangeFromPoint(50,0).startOffset is 24
Below first column:
FAIL document.caretRangeFromPoint(450,100).startOffset should be 12. Was 8.
FAIL document.caretRangeFromPoint(350,100).startOffset should be 12. Was 8.
FAIL document.caretRangeFromPoint(291,100).startOffset should be 12. Was 8.
Below second column:
PASS document.caretRangeFromPoint(290,100).startOffset is 24
PASS document.caretRangeFromPoint(250,100).startOffset is 24
PASS document.caretRangeFromPoint(191,100).startOffset is 24
To the right of the first column:
PASS document.caretRangeFromPoint(400,10).startOffset is 0
PASS document.caretRangeFromPoint(400,30).startOffset is 0
FAIL document.caretRangeFromPoint(400,50).startOffset should be 4. Was 8.
PASS document.caretRangeFromPoint(400,70).startOffset is 8
FAIL document.caretRangeFromPoint(400,90).startOffset should be 12. Was 8.
To the left of the third column:
PASS document.caretRangeFromPoint(0,10).startOffset is 24
FAIL document.caretRangeFromPoint(0,30).startOffset should be 27. Was 35.
FAIL document.caretRangeFromPoint(0,50).startOffset should be 31. Was 35.
FAIL document.caretRangeFromPoint(0,70).startOffset should be 35. Was 32.
In the gap between the first and second columns:
FAIL document.caretRangeFromPoint(291,20).startOffset should be 3. Was 7.
PASS document.caretRangeFromPoint(291,40).startOffset is 7
FAIL document.caretRangeFromPoint(291,60).startOffset should be 11. Was 15.
FAIL document.caretRangeFromPoint(291,69).startOffset should be 11. Was 8.
FAIL document.caretRangeFromPoint(290,20).startOffset should be 12. Was 8.
PASS document.caretRangeFromPoint(290,40).startOffset is 16
FAIL document.caretRangeFromPoint(290,60).startOffset should be 20. Was 24.
FAIL document.caretRangeFromPoint(290,69).startOffset should be 20. Was 24.
In the gap between the second and third columns:
PASS document.caretRangeFromPoint(191,20).startOffset is 15
FAIL document.caretRangeFromPoint(191,40).startOffset should be 19. Was 23.
FAIL document.caretRangeFromPoint(191,60).startOffset should be 23. Was 31.
FAIL document.caretRangeFromPoint(191,69).startOffset should be 23. Was 24.
PASS document.caretRangeFromPoint(190,20).startOffset is 24
FAIL document.caretRangeFromPoint(190,40).startOffset should be 28. Was 32.
PASS document.caretRangeFromPoint(190,60).startOffset is 32
PASS document.caretRangeFromPoint(190,69).startOffset is 32
PASS successfullyParsed is true

TEST COMPLETE
xxx xxx xxx xxx xxx xxx xxx xxx xxx
